backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.12",
    "bs4>=0.0.2",
    "fastapi>=0.115.12",
    "fastmcp>=2.8.0",
    "google-genai>=1.20.0",
    "mcp[cli]>=1.9.4",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.9.1",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "sentence-transformers>=4.1.0",
    "uvicorn>=0.34.3",
]
